requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
ollama==0.1.7
diffusers>=0.18.0
transformers>=4.30.0
accelerate>=0.20.0
torch
torchvision
torchaudio
numpy
pymongo==4.6.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
